requirements.txt
streamlit
python-dotenv
groq
langchain
langchain-community
faiss-cpu 
sentence-transformers
pypdf
numpy
tqdm

